requirements.txt
opencv-python
numpy
pyyaml
opencv-python-headless
python3-pyscard